requirements.txt
alembic==1.7.6
aniso8601==7.0.0
blinker==1.4
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
Flask-GraphQL==2.0.1
Flask-GraphQL-Auth==1.3.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphene-sqlalchemy-filter==1.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
promise==2.3
PyJWT==2.0.1
Rx==1.6.1
sentry-sdk==1.5.5
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.3
